Move all config-related code to separate file
